refactor: improve naming and project folders
