+++ package.json
@@ -45,7 +45,7 @@
     "@mui/x-date-pickers": "^8.3.0",
     "@reduxjs/toolkit": "^2.8.1",
     "dayjs": "^1.11.13",
-    "motion": "^12.12.1",
+    "motion": "^12.18.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.56.3",
