+++ src/pages/home/Home.tsx
@@ -4,7 +4,11 @@ import Typography from '@mui/material/Typography';
 import { motion } from 'motion/react';
 const Home = () => {
   return (
-    <Box animate={{ scale: 1 }} component={motion.div} initial={{ scale: 0 }}>
+    <Box
+      animate={{ opacity: 1 }}
+      component={motion.div}
+      initial={{ opacity: 0 }}
+    >
       <Typography
         component="h1"
         sx={{ textAlign: 'center' }}
