+++ src/pages/login-register/login/Login.tsx
@@ -96,7 +96,11 @@ const Login = () => {
   };
 
   return (
-    <Box animate={{ scale: 1 }} component={motion.div} initial={{ scale: 0 }}>
+    <Box
+      animate={{ opacity: 1 }}
+      component={motion.div}
+      initial={{ opacity: 0 }}
+    >
       {isLoading ? <Loading /> : null}
 
       <Typography
