--- src/pages/shop/cards/MockProducts.ts
@@ -1,95 +1,0 @@
-const MockProducts = [
-  {
-    id: 'd83faba7-99f2-4b13-8459-1a6228da3751',
-    categoryId: 'fef70e14-0c1f-4a51-8f6b-c61b10eca4b3',
-    name: 'Snake Plant',
-    slug: 'snake-plant',
-    description:
-      'A hardy succulent with upright, sword-like leaves. Snake Plants tolerate low light and require minimal care, making them one of the best choices for beginners. They improve air quality by filtering toxins. Their vertical growth makes them suitable for narrow spaces or modern interiors. Extremely resilient and adaptable, these plants thrive in almost any home setting.',
-    price: 0,
-    currencyCode: 'BYN',
-    images: [],
-  },
-  {
-    id: 'd83faba7-99f2-4b13-8459-1a6228da3752',
-    categoryId: 'fef70e14-0c1f-4a51-8f6b-c61b10eca4b3',
-    name: 'Snake Plant',
-    slug: 'snake-plant',
-    description:
-      'A hardy succulent with upright, sword-like leaves. Snake Plants tolerate low light and require minimal care, making them one of the best choices for beginners. They improve air quality by filtering toxins. Their vertical growth makes them suitable for narrow spaces or modern interiors. Extremely resilient and adaptable, these plants thrive in almost any home setting.',
-    price: 7500,
-    currencyCode: 'BYN',
-    images: [
-      {
-        url: 'https://images.squarespace-cdn.com/content/v1/6009d1db88b4396dd2d3d082/1645282569007-TSUBP7910E8FAEQEQ4BZ/86ACE3A7-8DEB-4E87-8EC4-4563F4FF5479.jpeg?format=1000w',
-        label: 'snake plant',
-        dimensions: {
-          w: 1000,
-          h: 1333,
-        },
-      },
-      {
-        url: 'https://www.livelyroot.com/cdn/shop/products/Snakes_Grower_4.jpg?v=1659635750&width=730',
-        dimensions: {
-          w: 730,
-          h: 913,
-        },
-      },
-    ],
-  },
-  {
-    id: 'd83faba7-99f2-4b13-8459-1a6228da3753',
-    categoryId: 'fef70e14-0c1f-4a51-8f6b-c61b10eca4b3',
-    name: 'Snake Plant',
-    slug: 'snake-plant',
-    description:
-      'A hardy succulent with upright, sword-like leaves. Snake Plants tolerate low light and require minimal care, making them one of the best choices for beginners. They improve air quality by filtering toxins. Their vertical growth makes them suitable for narrow spaces or modern interiors. Extremely resilient and adaptable, these plants thrive in almost any home setting.',
-    price: 7500,
-    currencyCode: 'BYN',
-    images: [
-      {
-        url: 'https://images.squarespace-cdn.com/content/v1/6009d1db88b4396dd2d3d082/1645282569007-TSUBP7910E8FAEQEQ4BZ/86ACE3A7-8DEB-4E87-8EC4-4563F4FF5479.jpeg?format=1000w',
-        label: 'snake plant',
-        dimensions: {
-          w: 1000,
-          h: 1333,
-        },
-      },
-      {
-        url: 'https://www.livelyroot.com/cdn/shop/products/Snakes_Grower_4.jpg?v=1659635750&width=730',
-        dimensions: {
-          w: 730,
-          h: 913,
-        },
-      },
-    ],
-  },
-  {
-    id: 'd83faba7-99f2-4b13-8459-1a6228da3754',
-    categoryId: 'fef70e14-0c1f-4a51-8f6b-c61b10eca4b3',
-    name: 'Snake Plant',
-    slug: 'snake-plant',
-    description:
-      'A hardy succulent with upright, sword-like leaves. Snake Plants tolerate low light and require minimal care, making them one of the best choices for beginners. They improve air quality by filtering toxins. Their vertical growth makes them suitable for narrow spaces or modern interiors. Extremely resilient and adaptable, these plants thrive in almost any home setting.',
-    price: 7590,
-    currencyCode: 'BYN',
-    images: [
-      {
-        url: 'https://images.squarespace-cdn.com/content/v1/6009d1db88b4396dd2d3d082/1645282569007-TSUBP7910E8FAEQEQ4BZ/86ACE3A7-8DEB-4E87-8EC4-4563F4FF5479.jpeg?format=1000w',
-        label: 'snake plant',
-        dimensions: {
-          w: 1000,
-          h: 1333,
-        },
-      },
-      {
-        url: 'https://www.livelyroot.com/cdn/shop/products/Snakes_Grower_4.jpg?v=1659635750&width=730',
-        dimensions: {
-          w: 730,
-          h: 913,
-        },
-      },
-    ],
-  },
-];
-export default MockProducts;
